+++ src/main/java/com/example/supplychainx/service_approvisionnement/service/SupplierService.java
@@ -2,26 +2,24 @@ package com.example.supplychainx.service_approvisionnement.service;
 
 import com.example.supplychainx.service_approvisionnement.dto.Supplier.SupplierRequestDTO;
 import com.example.supplychainx.service_approvisionnement.dto.Supplier.SupplierResponseDTO;
+import com.example.supplychainx.service_approvisionnement.exceptions.BusinessException;
 import com.example.supplychainx.service_approvisionnement.exceptions.SupplierNotFoundException;
 import com.example.supplychainx.service_approvisionnement.mapper.SupplierMapper;
 import com.example.supplychainx.service_approvisionnement.model.Supplier;
 import com.example.supplychainx.service_approvisionnement.repository.SupplierRepository;
+import com.example.supplychainx.service_approvisionnement.repository.SupplyOrderRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-
 import java.util.List;
 import java.util.stream.Collectors;
 
-
 @Service
+@RequiredArgsConstructor
 public class SupplierService {
     private final SupplierRepository supplierRepository;
     private final SupplierMapper supplierMapper;
+    private final SupplyOrderRepository supplyOrderRepository;
 
-    public SupplierService(SupplierRepository supplierRepository, SupplierMapper supplierMapper) {
-        this.supplierRepository = supplierRepository;
-        this.supplierMapper = supplierMapper;
-    }
 
     public SupplierResponseDTO createSupplier(SupplierRequestDTO dto) {
         Supplier supplier = supplierMapper.toEntity(dto);
@@ -56,6 +54,9 @@ public class SupplierService {
     public void deleteSupplier(Long id) {
         if (!supplierRepository.existsById(id)) {
             throw new SupplierNotFoundException("Fournisseur introuvable avec l'id : " + id);
+        }
+        if (supplyOrderRepository.existsBySupplierId(id)) {
+            throw new BusinessException("Impossible de supprimer le fournisseur. Il est associé à une ou plusieurs commandes.");
         }
         supplierRepository.deleteById(id);
     }
